split generateCatalogIndex.ts into patchNetworkRequests.ts
